Prune ALL test block papers from other partitions when making triplets
As in, all the papers from the test blocks, not just the ones that get sampled
into the test triplets.  This is because we might want to run full S2AND later
with a model that was trained on the triplets, so if a test-block paper made it
into the triplet train set we'd have a problem.